+++ testing/integration/src/lib.rs
@@ -1,0 +1,4 @@
+mod builtin;
+mod dummy;
+pub mod error;
+pub mod tester;